add the mvc pattern implementation demo
We are going to create a Student object acting as a model.StudentView
will be a view class which can print student details on console and
StudentController is the controller class responsible to store data in
Student object and update view StudentView accordingly.

MVCPatternDemo, our demo class, will use StudentController to
demonstrate use of MVC pattern.